First Commit
Only very basic things are done such as start menu, and game area with score at the bottom. The ADC value where the score gets its value from is temporary and is for debugging purposes.

Project name is SpaceInvaders but I do not know how to change on Keil, we should just keep it like that as it doesn't really matter too much.